refactor: 모델에서 승리 결과 추출 로직 분리
